Refactor the logic to find the file path and root path argument into a trait and update all commands to use the trait insteat. (#59)
